correctly implementing messages of type "error"
  Also: deferring variable validation to only happen during execution.
During client message parsing, it's not "our" job to validate if the
variables are there or not. This is already done by the executor later
and it provides a more adequate error message (actually, an error object).
However, I also found out that FSharp.Data.GraphQL has a bug in that
validation error is incorrectly treated as a fully unexpected scenario
and the resulting GraphQL message is simply an Exception.ToString() containing
the whole stack trace. This is a bug in this library and will have to be
fixed. So far, nothing to be done in graphql-ws.fsharp itself.