incoporated MessageMapping into the RawMessageConverter
 I was originally thinking about having a dumbed down converter which
was easy to test, but as a matter of fact it makes no sense to pass
around an object which is not completely deserialized.
 The undeserialized payload is now deliberate, since it's now up to the
client to define how to deserialize his own custom payload. The
possibility to configure this deserialization (which will happen
actually during the handling of a message like "connection_init" or
"pong") will come next.